% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_clustering_accuracy}
\alias{calculate_clustering_accuracy}
\title{Calculate clustering accuracy using best label matching}
\usage{
calculate_clustering_accuracy(true_labels, pred_labels)
}
\arguments{
\item{true_labels}{vector of true cluster labels}

\item{pred_labels}{vector of predicted cluster labels}
}
\value{
numeric accuracy score
}
\description{
Calculate clustering accuracy using best label matching
}
