% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_dist_matrix}
\alias{compute_dist_matrix}
\title{Compute pairwise absolute-distance matrix from a numeric vector}
\usage{
compute_dist_matrix(x)
}
\description{
Compute pairwise absolute-distance matrix from a numeric vector
}
