% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score_clustering.R
\name{plot.ps_clustering}
\alias{plot.ps_clustering}
\title{Plot method for ps_clustering objects}
\usage{
\method{plot}{ps_clustering}(x, which = 1:4, ...)
}
\arguments{
\item{x}{ps_clustering object}

\item{which}{integer vector specifying which plots to create (1-4)}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Plot method for ps_clustering objects
}
