% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score_clustering.R
\name{propensity_score_clustering}
\alias{propensity_score_clustering}
\title{Enhanced propensity score clustering}
\usage{
propensity_score_clustering(
  data,
  rhs_formula,
  k = min(100, floor(sqrt(nrow(data)))),
  max_groups = floor(sqrt(nrow(data))),
  min_improve = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{data}{data.frame of covariates}

\item{rhs_formula}{one-sided formula for PS (e.g. ~ X1 + X2 + X3)}

\item{k}{number of random-matchit iterations}

\item{max_groups}{maximum initial clusters (default = sqrt(n))}

\item{min_improve}{early stopping threshold on quality drop}

\item{verbose}{logical, whether to print progress}
}
\value{
ps_clustering object with clustering results
}
\description{
Enhanced propensity score clustering
}
