% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_initial_anchors}
\alias{select_initial_anchors}
\title{Select initial anchors via farthest-first traversal on mean PS}
\usage{
select_initial_anchors(meanPS, g_max)
}
\arguments{
\item{meanPS}{numeric vector of length n}

\item{g_max}{number of anchors to select}
}
\value{
integer vector of anchor indices
}
\description{
Select initial anchors via farthest-first traversal on mean PS
}
\keyword{internal}
