% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_dist_matrix}
\alias{compute_dist_matrix}
\title{Compute pairwise absolute-distance matrix from a numeric vector}
\usage{
compute_dist_matrix(x)
}
\arguments{
\item{x}{A numeric vector of length \eqn{n}, e.g. mean propensity scores.}
}
\value{
A numeric \eqn{n \times n} matrix where entry \eqn{[i,j]} is the Manhattan
distance \eqn{|x[i] - x[j]|}.
}
\description{
Compute pairwise absolute-distance matrix from a numeric vector
}
\keyword{internal}
