% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_plot_data}
\alias{create_plot_data}
\title{Create plot data for visualization (internal function)}
\usage{
create_plot_data(data, rhs_formula, cluster_labels, meanPS)
}
\description{
Create plot data for visualization (internal function)
}
\keyword{internal}
