% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score_clustering.R
\name{print.ps_clustering}
\alias{print.ps_clustering}
\title{Print method for ps_clustering objects}
\usage{
\method{print}{ps_clustering}(x, ...)
}
\arguments{
\item{x}{ps_clustering object}

\item{...}{additional arguments (unused)}
}
\description{
Print method for ps_clustering objects
}
